requirements.txt
streamlit 
pandasai
python-dotenv
bs4 
BeautifulSoup4
sqlalchemy
yfinance